// src/shared/enums/relationship.enum.ts
export enum RelationshipEnum {
  single = 'single',
  in_a_relationship = 'in_a_relationship',
  engaged = 'engaged',
  married = 'married',
  in_a_civil_union = 'in_a_civil_union',
  in_an_open_relationship = 'in_an_open_relationship',
  its_complicated = 'its_complicated',
  separated = 'separated',
  divorced = 'divorced',
  widowed = 'widowed',
}